requirements.txt
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/altgraph-0.17.2-py2.py3-none-any.whl
asgiref==3.5.2
big-O==0.10.2
certifi==2022.9.24
charset-normalizer==2.1.1
coverage==6.5.0
Django==4.1.3
django-crispy-forms==1.14.0
django-environ==0.9.0
django-watson==1.6.3
flake8==6.0.0
flake8-isort==6.0.0
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/future-0.18.2-py3-none-any.whl
idna==3.4
isort==5.11.4
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/macholib-1.15.2-py2.py3-none-any.whl
mccabe==0.7.0
numpy==1.24.1
Pillow==9.3.0
psycopg2==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
requests==2.28.1
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/six-1.15.0-py2.py3-none-any.whl
sqlparse==0.4.3
stripe==5.1.1
urllib3==1.26.12
